+++ packages/commonwealth/client/scripts/views/components/MarkdownEditor/customLinkDialog/CustomLinkDialogForDesktop.tsx
@@ -1,7 +1,7 @@
 import { linkDialogState$, useCellValues } from 'commonwealth-mdxeditor';
 import React, { useCallback, useEffect, useMemo } from 'react';
-import { CustomLinkEdit } from 'views/components/MarkdownEditor/CustomLinkEdit';
-import { CustomLinkPreview } from 'views/components/MarkdownEditor/CustomLinkPreview';
+import { CustomLinkEdit } from 'views/components/MarkdownEditor/customLinkDialog/CustomLinkEdit';
+import { CustomLinkPreview } from 'views/components/MarkdownEditor/customLinkDialog/CustomLinkPreview';
 import CWPopover, {
   usePopover,
 } from 'views/components/component_kit/new_designs/CWPopover';
