+++ packages/commonwealth/client/scripts/views/components/MarkdownEditor/customLinkDialog/CustomLinkDialogForMobile.tsx
@@ -1,7 +1,7 @@
 import { linkDialogState$, useCellValues } from 'commonwealth-mdxeditor';
 import React from 'react';
-import { CustomLinkEdit } from 'views/components/MarkdownEditor/CustomLinkEdit';
-import { CustomLinkPreview } from 'views/components/MarkdownEditor/CustomLinkPreview';
+import { CustomLinkEdit } from 'views/components/MarkdownEditor/customLinkDialog/CustomLinkEdit';
+import { CustomLinkPreview } from 'views/components/MarkdownEditor/customLinkDialog/CustomLinkPreview';
 
 /**
  * We need to use a popover on desktop and no container no mobile.
